Src/Main_Algorithm/requirements.txt
wheel
numpy
datasets
transformers
pydub==0.25.1
librosa==0.10.1
ffmpeg-python==0.2.0
pyannote.audio==3.1.1



